+++ stripe_terminal/android/src/main/kotlin/mek/stripeterminal/plugin/ReaderReconnectionDelegatePlugin.kt
@@ -3,11 +3,11 @@ package mek.stripeterminal.plugin
 import com.stripe.stripeterminal.external.callable.Cancelable
 import com.stripe.stripeterminal.external.callable.ReaderReconnectionListener
 import com.stripe.stripeterminal.external.models.Reader
-import mek.stripeterminal.api.StripeTerminalHandlersApi
+import mek.stripeterminal.api.TerminalHandlersApi
 import mek.stripeterminal.api.toApi
 import mek.stripeterminal.runOnMainThread
 
-class ReaderReconnectionListenerPlugin(private val _handlers: StripeTerminalHandlersApi) :
+class ReaderReconnectionListenerPlugin(private val _handlers: TerminalHandlersApi) :
     ReaderReconnectionListener {
     var cancelReconnect: Cancelable? = null